Restructured Tunnel modules
